requirements.txt
django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
pillow==10.4.0
celery==5.4.0
redis==5.0.8
requests==2.32.3
mysqlclient==2.2.4
